Moved code related to the outgoing pipeline into the same folder
