+++ src/NServiceBus.Core.Tests/Routing/DetermineRouteForSendBehaviorTests.cs
@@ -2,7 +2,6 @@
 {
     using System;
     using NServiceBus.OutgoingPipeline;
-    using NServiceBus.OutgoingPipeline.Send;
     using NServiceBus.Pipeline.Contexts;
     using NServiceBus.Routing;
     using NUnit.Framework;
