+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -154,23 +154,24 @@
     <Compile Include="IBusExtensions.cs" />
     <Compile Include="IDataBusProperty.cs" />
     <Compile Include="ISendOnlyBusExtensions.cs" />
-    <Compile Include="MessageMutator\IMutateOutgoingTransportMessages.cs" />
-    <Compile Include="MessageMutator\MutateOutgoingMessagesContext.cs" />
-    <Compile Include="MessageMutator\MutateOutgoingTransportMessagesContext.cs" />
+    <Compile Include="OutgoingPipeline\Mutators\IMutateOutgoingTransportMessages.cs" />
+    <Compile Include="OutgoingPipeline\Mutators\MutateOutgoingMessagesContext.cs" />
+    <Compile Include="OutgoingPipeline\Mutators\MutateOutgoingTransportMessagesContext.cs" />
     <Compile Include="Extensibility\OptionExtensionContext.cs" />
     <Compile Include="OutgoingPipeline\OutgoingLogicalMessage.cs" />
     <Compile Include="OutgoingPipeline\OutgoingLogicalMessageContextExtensions.cs" />
-    <Compile Include="OutgoingPipeline\Publish\PublishToOutgoingContextConnector.cs" />
-    <Compile Include="OutgoingPipeline\Send\OutgoingSendContext.cs" />
-    <Compile Include="OutgoingPipeline\Send\SendToOutgoingContextConnector.cs" />
-    <Compile Include="OutgoingPipeline\Reply\OutgoingReplyContext.cs" />
-    <Compile Include="OutgoingPipeline\Reply\ReplyToOutgoingContextConnector.cs" />
+    <Compile Include="OutgoingPipeline\OutgoingPipelineFeature.cs" />
+    <Compile Include="OutgoingPipeline\PublishToOutgoingContextConnector.cs" />
+    <Compile Include="OutgoingPipeline\OutgoingSendContext.cs" />
+    <Compile Include="OutgoingPipeline\SendToOutgoingContextConnector.cs" />
+    <Compile Include="OutgoingPipeline\OutgoingReplyContext.cs" />
+    <Compile Include="OutgoingPipeline\ReplyToOutgoingContextConnector.cs" />
     <Compile Include="Performance\MessageDurability\DetermineMessageDurabilityBehavior.cs" />
     <Compile Include="Performance\MessageDurability\DurableMessagesConventionExtensions.cs" />
     <Compile Include="Performance\MessageDurability\MessageDurabilityFeature.cs" />
     <Compile Include="Extensibility\ContextBag.cs" />
     <Compile Include="Pipeline\Contexts\LogicalMessagesContextExtensions.cs" />
-    <Compile Include="OutgoingPipeline\Publish\OutgoingPublishContext.cs" />
+    <Compile Include="OutgoingPipeline\OutgoingPublishContext.cs" />
     <Compile Include="Pipeline\Contexts\TransportMessageContextExtensions.cs" />
     <Compile Include="Pipeline\IPipelineTerminator.cs" />
     <Compile Include="Pipeline\PipelineDiagnostics.cs" />
@@ -212,6 +213,7 @@
     <Compile Include="Routing\StorageDrivenPublishing\SubscribersForEvent.cs" />
     <Compile Include="Routing\RoutingAdapter.cs" />
     <Compile Include="Routing\RoutingContextExtensions.cs" />
+    <Compile Include="Sagas\AutoCorrelationFeature.cs" />
     <Compile Include="Sagas\PopulateAutoCorrelationHeadersForRepliesBehavior.cs" />
     <Compile Include="Satellites\IAdvancedSatellite.cs" />
     <Compile Include="Satellites\ISatellite.cs" />
@@ -246,7 +248,7 @@
     <Compile Include="Transports\Msmq\MsmqReceiveWithNoTransactionBehavior.cs" />
     <Compile Include="Transports\OutgoingMessage.cs" />
     <Compile Include="Transports\ReceiveBehavior.cs" />
-    <Compile Include="Pipeline\Contexts\PhysicalOutgoingContext.cs" />
+    <Compile Include="OutgoingPipeline\PhysicalOutgoingContextStageBehavior.cs" />
     <Compile Include="Transports\IncomingMessage.cs" />
     <Compile Include="Transports\TransportPublishOptions.cs" />
     <Compile Include="Transports\DispatchOptions.cs" />
@@ -432,9 +434,9 @@
     <Compile Include="Logging\RollingLogger.cs" />
     <Compile Include="Conventions.cs" />
     <Compile Include="MessageIntentEnum.cs" />
-    <Compile Include="MessageMutator\IMutateIncomingMessages.cs" />
-    <Compile Include="MessageMutator\IMutateIncomingTransportMessages.cs" />
-    <Compile Include="MessageMutator\IMutateOutgoingMessages.cs" />
+    <Compile Include="OutgoingPipeline\Mutators\IMutateIncomingMessages.cs" />
+    <Compile Include="OutgoingPipeline\Mutators\IMutateIncomingTransportMessages.cs" />
+    <Compile Include="OutgoingPipeline\Mutators\IMutateOutgoingMessages.cs" />
     <Compile Include="Reliability\Outbox\IOutboxStorage.cs" />
     <Compile Include="Reliability\Outbox\Outbox.cs" />
     <Compile Include="Reliability\Outbox\OutboxConfigExtensions.cs" />
@@ -455,8 +457,8 @@
     <Compile Include="Persistence\PersistenceConfig.cs" />
     <Compile Include="Pipeline\BehaviorContext.cs" />
     <Compile Include="Pipeline\BehaviorContextStacker.cs" />
-    <Compile Include="MessageMutator\ApplyIncomingMessageMutatorsBehavior.cs" />
-    <Compile Include="MessageMutator\ApplyIncomingTransportMessageMutatorsBehavior.cs" />
+    <Compile Include="OutgoingPipeline\Mutators\ApplyIncomingMessageMutatorsBehavior.cs" />
+    <Compile Include="OutgoingPipeline\Mutators\ApplyIncomingTransportMessageMutatorsBehavior.cs" />
     <Compile Include="Pipeline\StepRegistrationsCoordinator.cs" />
     <Compile Include="Pipeline\BehaviorTypeChecker.cs" />
     <Compile Include="Pipeline\Dependency.cs" />
@@ -497,8 +499,8 @@
     <Compile Include="Unicast\Behaviors\MessageHandler.cs" />
     <Compile Include="Unicast\Behaviors\SerializeMessagesBehavior.cs" />
     <Compile Include="TransportDispatch\DispatchMessageToTransportConnector.cs" />
-    <Compile Include="MessageMutator\MutateOutgoingMessageBehavior.cs" />
-    <Compile Include="MessageMutator\MutateOutgoingPhysicalMessageBehavior.cs" />
+    <Compile Include="OutgoingPipeline\Mutators\MutateOutgoingMessageBehavior.cs" />
+    <Compile Include="OutgoingPipeline\Mutators\MutateOutgoingPhysicalMessageBehavior.cs" />
     <Compile Include="Pipeline\Contexts\LogicalMessagesProcessingStageBehavior.cs" />
     <Compile Include="Pipeline\PipelineBase.cs" />
     <Compile Include="Pipeline\Contexts\RootContext.cs" />
