+++ src/NServiceBus.Core/OutgoingPipeline/Mutators/IMutateIncomingMessages.cs
