+++ src/NServiceBus.Core/OutgoingPipeline/Mutators/IMutateOutgoingMessages.cs
