+++ src/NServiceBus.Core/OutgoingPipeline/Mutators/MutateOutgoingMessageBehavior.cs
