+++ src/NServiceBus.Core/OutgoingPipeline/PhysicalOutgoingContextStageBehavior.cs
@@ -1,5 +1,8 @@
-﻿namespace NServiceBus.Pipeline.Contexts
+﻿namespace NServiceBus.OutgoingPipeline
 {
+    using NServiceBus.Pipeline;
+    using NServiceBus.Pipeline.Contexts;
+
     /// <summary>
     /// Represent the part of the outgoing pipeline where the message has been serialized to a byte[]
     /// </summary>
@@ -22,7 +25,7 @@
 
 
             /// <summary>
-            /// 
+            /// The serialized body of the outgoing message
             /// </summary>
             /// <summary>
             /// A <see cref="byte"/> array containing the serialized contents of the outgoing message.
