+++ src/NServiceBus.Core/OutgoingPipeline/PublishToOutgoingContextConnector.cs
@@ -1,7 +1,7 @@
 namespace NServiceBus
 {
     using System;
-    using NServiceBus.OutgoingPipeline.Publish;
+    using NServiceBus.OutgoingPipeline;
     using NServiceBus.Pipeline;
     using NServiceBus.Pipeline.Contexts;
