+++ src/NServiceBus.Core/OutgoingPipeline/SendToOutgoingContextConnector.cs
@@ -1,7 +1,7 @@
 namespace NServiceBus
 {
     using System;
-    using NServiceBus.OutgoingPipeline.Send;
+    using NServiceBus.OutgoingPipeline;
     using NServiceBus.Pipeline;
     using NServiceBus.Pipeline.Contexts;
