+++ src/NServiceBus.Core/Routing/DetermineRouteForPublishBehavior.cs
@@ -2,7 +2,6 @@ namespace NServiceBus
 {
     using System;
     using NServiceBus.OutgoingPipeline;
-    using NServiceBus.OutgoingPipeline.Publish;
     using NServiceBus.Pipeline;
     using NServiceBus.Routing;
     using NServiceBus.TransportDispatch;
